+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Strip markdown code fences from rewrite results via new `rewrite/contentReceived` `replace` content type. #336
+
 ## 0.109.0
 
 - Improve /login text.